+++ msdos/src/fn_fuji/fuji_qrcode_output.c
@@ -1,0 +1,9 @@
+#include <stdbool.h>
+#include <stdint.h>
+#include "fujinet-fuji.h"
+#include "fujinet-fuji-msdos.h"
+
+bool fuji_qrcode_output(char *s, uint16_t len)
+{
+    return int_f5_read(0x70, FUJICMD_QRCODE_OUTPUT, len & 0xFF, len >> 8, s, len) == 'C';
+}